+++ R/ed_holiday.R
@@ -1,0 +1,5 @@
+# http://www.gov.cn/zfwj/bgtfd.htm
+# http://www.gov.cn/zhengce/xxgkzl.htm
+# http://sousuo.gov.cn/s.htm?t=paper&advance=true&sort=&title=年节假日安排&content=&puborg=&pcodeJiguan=&pcodeYear=&pcodeNum=&timetype=timeqb&mintime=&maxtime=&filetype=&childtype=&subchildtype=
+# type: 1-workday; 3-weekend; 4-holiday;
+